Bring in lots of Qcow2 code
Brings in a TON of QCOW2 code right from RSD. I decided to go a
different approach with this and just bring a bunch of stuff in and then
eventually just take out what I don't need.

A lot of the code was heavily tied up into the QMP bits, which we don't
care about and don't need. I commented out what I needed to in order for
things to compile but there is going to be more gutting that needs to
happen.

Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>