intermediary commit. not done trying to get things compiled
Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>

this somehow compiles with everything... Ill take it

Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>

Trying to get this to boot

Currently it tries to boot the image but it just results in a blank
screen. I think this _may_ have to do with some functions that are
`todo!()` but I'm not sure... those don't have any implementation in
crosvm so I'm not sure if there's really a good way to go about this.

We shall see...

Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>

this might be working on linux. definitely a mess of a code change and some of the stuff that was added will be linux only and needs to be moved around/flagged as such

Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>