Trying to get this to boot
Currently it tries to boot the image but it just results in a blank
screen. I think this _may_ have to do with some functions that are
`todo!()` but I'm not sure... those don't have any implementation in
crosvm so I'm not sure if there's really a good way to go about this.

We shall see...

Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>